Move error type.
